Time: 4 ms (25.54%), Space: 21.4 MB (9.17%) - LeetHub

+++ 0871-minimum-number-of-refueling-stops/0871-minimum-number-of-refueling-stops.cpp
@@ -1,15 +1,17 @@
 class Solution {
 public:
     int minRefuelStops(int target, int startFuel, vector<vector<int>>& stations) {
+        //step 1.add the starting in stations
         stations.push_back({target,0});
+        //step 2.create priority queue
         priority_queue<int> pq;
-        long long fuel= startFuel;
+        long long fuel=startFuel;
         long long prev=0;
         int stops=0;
         for(auto station: stations){
-            long long position=station[0];
-            long long stationFuel=station[1];
-            long long distance= position-prev;
+            int position= station[0];
+            int stationFuel=station[1];
+            int distance= position- prev;
             fuel-=distance;
             while(fuel<0 && !pq.empty()){
                 fuel+=pq.top();
